# app/Virtual/UpdateRoleRequest.php
<?php

namespace App\Virtual;

/**
 * @OA\Schema(
 *      title="Update Role request",
 *      description="Update Role request body data",
 *      type="object",
 *      required={"name"}
 * )
 */

class UpdateRoleRequest
{
    /**
     * @OA\Property(
     *      title="name",
     *      description="Name of the new role",
     *      example="A nice role"
     * )
     *
     * @var string
     */
    public $name;

    /**
     * @OA\Property(
     *      title="description",
     *      description="Description of the new role",
     *      example="This is new role's description"
     * )
     *
     * @var string
     */
    public $description;

    /**
     * @OA\Property(
     *      title="author_id",
     *      description="Author's id of the new role",
     *      format="int64",
     *      example=1
     * )
     *
     * @var integer
     */
    public $author_id;
}